asklvkaszus-backend/requirements.txt
bcrypt==4.2.0
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
Deprecated==1.2.14
Flask==3.0.3
Flask-Cors==4.0.1
Flask-Limiter==3.8.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.3
gunicorn==23.0.0
idna==3.7
importlib_resources==6.4.4
itsdangerous==2.2.0
Jinja2==3.1.4
limits==3.13.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mysql-connector-python==9.0.0
ordered-set==4.1.0
packaging==24.1
protobuf==5.27.3
Pygments==2.18.0
PyJWT==2.9.0
PyMySQL==1.1.1
PyYAML==6.0.2
redis==5.0.8
requests==2.32.3
rich==13.7.1
SQLAlchemy==2.0.32
typing_extensions==4.12.2
urllib3==2.2.2
uuid==1.30
Werkzeug==3.0.4
wheel==0.44.0
wrapt==1.16.0
WTForms==3.1.2
